+++ fe/src/hooks/useCheckAuth.js
@@ -1,4 +1,4 @@
-import { useContext, useEffect } from 'react';
+import { useEffect } from 'react';
 import { useHistory, useLocation } from 'react-router-dom';
 import { useUserDispatch } from '../context/index';
 import API from '../lib/api';
